Removes more f5-sdk references from f5 modules
